2016-11-27 Fred Gleason <fredg@paravelsystems.com> 	* Changed 'MAY_DELAY' to '60' in 'src/common/config.h'. 	* Fixed a bug in 'src'lwnetrouterd/lwnetrouterd.cpp' that caused 	calculation of delay interval top be incorrect for intervals 	greater than 15 seconds.

+++ src/lwnetrouterd/router_hpiaudio.cpp
@@ -29,7 +29,7 @@
 
 #include "router_hpiaudio.h"
 
-size_t SetpointBytes(uint32_t msec)
+size_t SetpointBytes(uint64_t msec)
 {
   return AUDIO_SAMPLE_RATE*AUDIO_CHANNELS*msec*sizeof(float)/1000;
 }
